Refactor away things from LootLockerConfig to more appropriate utilities
