+++ LootLockerSDK/Source/LootLockerSDK/Private/Utils/LootLockerUtilities.cpp
@@ -14,6 +14,19 @@ const TArray<FObfuscationDetails> UObfuscationSettings::FieldsToObfuscate =
     FObfuscationDetails(FString("token"), "*", 3, 3, true)
 };
 
+FString ULootLockerEnumUtils::GetEnum(const TCHAR* Enum, int32 EnumValue)
+{
+#if ENGINE_MAJOR_VERSION <= 4 && ENGINE_MINOR_VERSION <= 27
+    const UEnum* EnumPtr = FindObject<UEnum>(StaticClass()->GetOuter(), Enum, true);
+#else
+    const UEnum* EnumPtr = FindObject<UEnum>(StaticClass()->GetOuterUPackage(), Enum, true);
+#endif
+    if (!EnumPtr)
+        return NSLOCTEXT("Invalid", "Invalid", "Invalid").ToString();
+
+    return EnumPtr->GetDisplayNameTextByValue(EnumValue).ToString();
+}
+
 namespace LootLockerUtilities
 {
     FString AppendParameterToUrl(const FString& Url, const FString& Parameter)
