+++ LootLockerSDK/Source/LootLockerSDK/Public/GameAPI/LootLockerUserGeneratedContentRequestHandler.h
@@ -3,10 +3,7 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "UObject/NoExportTypes.h"
-#include "LootLockerConfig.h"
 #include "LootLockerResponse.h"
-#include "JsonObjectConverter.h"
 #include "LootLockerHttpClient.h"
 #include "LootLockerUserGeneratedContentRequestHandler.generated.h"
