+++ LootLockerSDK/Source/LootLockerSDK/Public/LootLockerConfig.h
@@ -3,69 +3,12 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "LootLockerResponse.h"
 #include "LootLockerConfig.generated.h"
-
-DECLARE_LOG_CATEGORY_EXTERN(LogLootLockerGameSDK, Log, All);
-
-UENUM(BlueprintType)
-enum class ELootLockerHTTPMethod : uint8
-{
-	GET = 0         UMETA(DisplayName = "GET"),
-	POST = 1        UMETA(DisplayName = "POST"),
-	DELETE = 2      UMETA(DisplayName = "DELETE"),
-	PUT = 3         UMETA(DisplayName = "PUT"),
-	HEAD = 4        UMETA(DisplayName = "HEAD"),
-	CREATE = 5      UMETA(DisplayName = "CREATE"),
-	OPTIONS = 6     UMETA(DisplayName = "OPTIONS"),
-	PATCH = 7       UMETA(DisplayName = "PATCH"),
-	UPLOAD = 8      UMETA(DisplayName = "UPLOAD")
-};
-
-UENUM(BlueprintType)
-enum class ELootLockerPlatformType : uint8
-{
-	Android = 0				UMETA(DisplayName = "Android"),
-	Ios = 1					UMETA(DisplayName = "Ios"),
-	Steam = 2				UMETA(DisplayName = "Steam"),
-	NintendoSwitch = 3		UMETA(DisplayName = "NintendoSwitch"),
-	PlayStationNetwork = 4  UMETA(DisplayName = "PlayStationNetwork"),
-	Xbox = 5				UMETA(DisplayName = "Xbox"),
-	UNUSED = 6				UMETA(DisplayName = "Unused")
-};
-
-USTRUCT(BlueprintType)
-struct FLootLockerEndPoints
-{
-	GENERATED_BODY()
-public:
-	UPROPERTY(BlueprintReadWrite, EditAnywhere, BlueprintReadWrite, Category = "LootLocker")
-	FString endpoint;
-	UPROPERTY(BlueprintReadWrite, EditAnywhere, BlueprintReadWrite, Category = "LootLocker")
-	ELootLockerHTTPMethod requestMethod = ELootLockerHTTPMethod::GET;
-};
-
-USTRUCT(BlueprintType)
-struct FLootLockerGetRequests
-{
-	GENERATED_BODY()
-	TArray<FStringFormatArg> args;
-};
-
-DECLARE_DYNAMIC_DELEGATE_OneParam(FResponseCallbackBP, FLootLockerResponse, Response);
-DECLARE_DYNAMIC_DELEGATE_OneParam(FLootLockerResponseCallback, FLootLockerResponse, Response);
-DECLARE_DELEGATE_OneParam(FResponseCallback, FLootLockerResponse);
 
 UCLASS(Config = Game, DefaultConfig, meta = (DisplayName = "LootLocker SDK Settings"))
 class LOOTLOCKERSDK_API ULootLockerConfig : public UObject
 {
 	GENERATED_BODY()
-public:
-	static FString GetEnum(const TCHAR* Enum, int32 EnumValue);
-    static FString GetRequestMethodString(ELootLockerHTTPMethod RequestMethod)
-    {
-        return GetEnum(TEXT("ELootLockerHTTPMethod"), static_cast<int32>(RequestMethod));
-    }
 public:
 	// API Key used to talk to LootLocker. The API key can be found in `Settings > API Keys` in the Web Console: https://console.lootlocker.com/settings/api-keys
 	UPROPERTY(Config, EditAnywhere, BlueprintReadWrite, Category = "LootLocker", Meta = (DisplayName = "LootLocker API Key"))
