+++ LootLockerSDK/Source/LootLockerSDK/Public/LootLockerPlatformManager.h
@@ -2,12 +2,21 @@
 
 #pragma once
 
-#include <stdexcept>
-
 #include "CoreMinimal.h"
-#include "LootLockerConfig.h"
 #include "LootLockerStateData.h"
 #include "LootLockerPlatformManager.generated.h"
+
+UENUM(BlueprintType)
+enum class ELootLockerPlatformType : uint8
+{
+	Android = 0				UMETA(DisplayName = "Android"),
+	Ios = 1					UMETA(DisplayName = "Ios"),
+	Steam = 2				UMETA(DisplayName = "Steam"),
+	NintendoSwitch = 3		UMETA(DisplayName = "NintendoSwitch"),
+	PlayStationNetwork = 4  UMETA(DisplayName = "PlayStationNetwork"),
+	Xbox = 5				UMETA(DisplayName = "Xbox"),
+	UNUSED = 6				UMETA(DisplayName = "Unused")
+};
 
 UENUM(BlueprintType)
 enum class ELootLockerPlatform : uint8
