+++ LootLockerSDK/Source/LootLockerSDK/Public/LootLockerSDK.h
@@ -4,8 +4,8 @@
 
 #include "CoreMinimal.h"
 #include "Modules/ModuleManager.h"
-#include "LootLockerConfig.h"
 
+DECLARE_LOG_CATEGORY_EXTERN(LogLootLockerGameSDK, Log, All);
 class FLootLockerSDKModule : public IModuleInterface
 {
 public:
